requirements.txt
# Core Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-JWT-Extended==4.5.3
Werkzeug==3.0.1

# Alternative: FastAPI (uncomment if using FastAPI instead of Flask)
# fastapi==0.109.0
# uvicorn==0.27.0
# python-multipart==0.0.6

# Database
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# AI/ML Libraries
openai==1.10.0
google-generativeai==0.3.2
anthropic==0.8.1
langchain==0.1.4
langchain-community==0.0.16

# IBM Watson
ibm-watson==7.0.1
ibm-cloud-sdk-core==3.18.2
ibm-watsonx-ai==0.2.0

# Machine Learning
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0
scipy==1.12.0
joblib==1.3.2

# Deep Learning (Optional)
tensorflow==2.15.0
torch==2.1.2
transformers==4.37.0

# NLP
nltk==3.8.1
spacy==3.7.2
sentence-transformers==2.3.1
textblob==0.17.1

# Data Processing
python-dotenv==1.0.1
pyyaml==6.0.1
jsonschema==4.20.0
marshmallow==3.20.1

# HTTP Requests
requests==2.31.0
httpx==0.26.0
aiohttp==3.9.1

# Task Queue & Scheduling
celery==5.3.6
redis==5.0.1
apscheduler==3.10.4

# Email
Flask-Mail==0.9.1
sendgrid==6.11.0

# Authentication & Security
bcrypt==4.1.2
PyJWT==2.8.0
cryptography==42.0.0
python-jose==3.3.0

# Validation
email-validator==2.1.0
phonenumbers==8.13.27
validators==0.22.0

# API Documentation
flasgger==0.9.7.1
flask-swagger-ui==4.11.1

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.23.3
requests-mock==1.11.0

# Code Quality
black==24.1.1
flake8==7.0.0
pylint==3.0.3
mypy==1.8.0

# Monitoring & Logging
python-json-logger==2.0.7
loguru==0.7.2

# Deployment
gunicorn==21.2.0
gevent==24.2.1
supervisor==4.2.5

# Utilities
click==8.1.7
tqdm==4.66.1
python-dateutil==2.8.2
pytz==2024.1
colorama==0.4.6

# Web Scraping (for job data collection)
beautifulsoup4==4.12.3
selenium==4.17.2
scrapy==2.11.0

# Data Visualization
matplotlib==3.8.2
seaborn==0.13.2
plotly==5.18.0

# Caching
Flask-Caching==2.1.0

# Rate Limiting
Flask-Limiter==3.5.1

# File Processing
openpyxl==3.1.2
python-docx==1.1.0
PyPDF2==3.0.1

# Environment
python-decouple==3.8